refactor(php-api-client-base): rename ApiClientConfiguration to ApiClientOptions
Remove the authenticator field from the options object in preparation
for making it a first-class ApiClient constructor argument. Rename the
test file to match the new class name and drop the authenticator-related
assertions.
